+++ src/views/GeekTalk/overwrite.css
@@ -1,0 +1,38 @@
+.el-form-item__label {
+    line-height: 20px;
+}
+.el-select {
+    width: 100%;
+}
+.el-tabs__header {
+    padding: 0;
+    background-color: #fff;
+    position: relative;
+    margin: 0 0 8px;
+    
+}
+.el-tabs_header, .el-tabs__nav-wrap, .el-tabs__nav-scroll {
+    height: 45px;
+    min-height: 45px;
+    max-height: 45px;
+}
+.el-tabs__nav {
+    height: 42px;
+    margin: 0 20px;
+}
+.el-card__body {
+    height: 100%;
+    min-height: 100%;
+    padding: 0;
+    display: flex;
+    flex-direction: column;
+}
+.el-avatar {
+    width: 45px;
+    height: 45px;
+    line-height: 45px;
+}
+
+.el-dialog__body {
+    padding: 10px 20px 20px 20px;
+}
